% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_aghq.R
\name{aghq_extract_draw_blocks}
\alias{aghq_extract_draw_blocks}
\title{Extract AGHQ draw blocks for theta and random effects}
\usage{
aghq_extract_draw_blocks(samps, coef_meta, time_varying_betas)
}
\description{
Extract AGHQ draw blocks for theta and random effects
}
\keyword{internal}
