% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{build_combined_mesh}
\alias{build_combined_mesh}
\title{(Internal) Build combined mesh from a list of sf polygons}
\usage{
build_combined_mesh(polygon_list, mesh_args, make_mesh)
}
\arguments{
\item{polygon_list}{List of 'sf' objects (same CRS).}

\item{mesh_args}{Passed to 'build_mesh()'.}
}
\value{
An 'inla.mesh' object or 'NULL' if 'make_mesh = FALSE'.
}
\description{
Just in case some maps have additional polygons outside the first extent
}
\keyword{internal}
