% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_coef_meta}
\alias{compute_coef_meta}
\title{Compute coefficient metadata (number of timepoints, covariates, and their names)}
\usage{
compute_coef_meta(data)
}
\arguments{
\item{data}{A disag_data_mmap object.}
}
\description{
Compute coefficient metadata (number of timepoints, covariates, and their names)
}
\keyword{internal}
