% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{disag_model_mmap}
\alias{disag_model_mmap}
\title{Fit a multi‑map disaggregation model (via AGHQ or TMB)}
\usage{
disag_model_mmap(
  data,
  priors = NULL,
  family = "poisson",
  link = "log",
  engine = c("AGHQ", "TMB"),
  k = 3,
  field = TRUE,
  iid = TRUE,
  silent = TRUE,
  starting_values = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{engine}{Character; either "AGHQ" or "TMB"}
}
\description{
Fit a multi‑map disaggregation model (via AGHQ or TMB)
}
