% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_helpers.R
\name{is_categorical_layer}
\alias{is_categorical_layer}
\title{Detect if a raster layer is categorical}
\usage{
is_categorical_layer(
  raster_layer,
  layer_name,
  categorical_baselines = NULL,
  max_categories = 10
)
}
\arguments{
\item{raster_layer}{A SpatRaster layer to check}

\item{layer_name}{The name of the layer}

\item{categorical_baselines}{Named list of categorical baselines from disag_data_mmap}

\item{max_categories}{Maximum number of unique values to consider categorical (default = 10)}
}
\value{
Logical indicating if the layer should be treated as categorical
}
\description{
Determines if a raster layer should be treated as categorical based on multiple criteria:
1. Explicit definition in categorical_covariate_baselines
2. Presence of defined levels in the raster
3. Small number of unique values
}
\keyword{internal}
