% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_aggregation_raster}
\alias{plot_aggregation_raster}
\title{Plot the offset raster}
\usage{
plot_aggregation_raster(disag_data, time = 1)
}
\arguments{
\item{disag_data}{A 'disag_data_mmap' object.}

\item{time}{Integer time-slice (default = 1).}
}
\value{
A ggplot2 object.
}
\description{
Draws the aggregation pixel values used in the fit
}
