% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_mesh}
\alias{plot_mesh}
\title{Plot the SPDE mesh with custom outer/inner boundaries}
\usage{
plot_mesh(
  disag_data,
  edge_col = "grey70",
  edge_size = 0.2,
  outer_col = "black",
  outer_size = 1,
  inner_col = "blue",
  inner_size = 1,
  node_col = "black",
  node_size = 0.5
)
}
\arguments{
\item{disag_data}{A 'disag_data_mmap' object.}

\item{edge_col}{Colour for internal mesh edges (default = "grey70").}

\item{edge_size}{Line width for those edges (default = 0.2).}

\item{outer_col}{Colour for the outer perimeter (default = "black").}

\item{outer_size}{Line width for the outer perimeter (default = 1).}

\item{inner_col}{Colour for any inner perimeter (default = "blue").}

\item{inner_size}{Line width for inner perimeter (default = 1).}

\item{node_col}{Colour for mesh nodes (default = "black").}

\item{node_size}{Size for mesh nodes (default = 0.5).}
}
\value{
A ggplot2 object.
}
\description{
Plot the SPDE mesh with custom outer/inner boundaries
}
