% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_polygons}
\alias{plot_polygons}
\title{Plot polygon response data}
\usage{
plot_polygons(disag_data, time = 1, show_title = TRUE)
}
\arguments{
\item{disag_data}{A 'disag_data_mmap' object.}

\item{time}{Integer index of time-slice to plot (default = 1).}

\item{show_title}{Logical; if TRUE (default), add a title "Response at time X".}
}
\value{
A ggplot2 object.
}
\description{
Draws the prepared polygons colored by the response variable, with an optional title.
}
