% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_mcmc.R
\name{predict.disag_model_mmap_mcmc}
\alias{predict.disag_model_mmap_mcmc}
\title{Predict for Multi-Map Disaggregation Model fit with MCMC}
\usage{
\method{predict}{disag_model_mmap_mcmc}(
  object,
  new_data = NULL,
  predict_iid = FALSE,
  N = 100,
  CI = 0.95,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{object}{A fitted \code{disag_model_mmap_mcmc} object.}

\item{new_data}{Optionally, a new SpatRaster (or list of them) for prediction.}

\item{predict_iid}{Logical. If TRUE, include the polygon IID effect in predictions.}

\item{N}{Number of posterior draws to use for uncertainty estimation.}

\item{CI}{Credible interval level (default 0.95).}

\item{verbose}{Logical; if TRUE, print runtime diagnostics.}

\item{...}{Further arguments.}
}
\description{
Predict for Multi-Map Disaggregation Model fit with MCMC
}
