% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_tmb.R
\name{predict.disag_model_mmap_tmb}
\alias{predict.disag_model_mmap_tmb}
\title{Predict for Multi-Map Disaggregation Model fit with TMB}
\usage{
\method{predict}{disag_model_mmap_tmb}(object, new_data = NULL, predict_iid = FALSE, N = 100, CI = 0.95, ...)
}
\arguments{
\item{object}{A fitted disag_model_mmap_tmb object.}

\item{new_data}{Optionally, a new SpatRaster (or list of them) for prediction.}

\item{predict_iid}{Logical. If TRUE, include the polygon iid effect in predictions.}

\item{N}{Number of Monte Carlo draws for uncertainty estimation.}

\item{CI}{Credible interval level (default 0.95).}

\item{...}{Further arguments.}
}
\description{
Predict for Multi-Map Disaggregation Model fit with TMB
}
