% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.disag_model_mmap_aghq}
\alias{print.disag_model_mmap_aghq}
\title{Summarize model fit for multi-map disaggregation model fit with AGHQ using AGHQ's builtin summary}
\usage{
\method{print}{disag_model_mmap_aghq}(object)
}
\arguments{
\item{object}{}
}
\description{
Summarize model fit for multi-map disaggregation model fit with AGHQ using AGHQ's builtin summary
}
