% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.disag_model_mmap_mcmc}
\alias{print.disag_model_mmap_mcmc}
\title{Print method for 'disag_model_mmap_mcmc' objects}
\usage{
\method{print}{disag_model_mmap_mcmc}(x, ...)
}
\arguments{
\item{x}{A 'disag_model_mmap_mcmc' object.}

\item{...}{Additional arguments (not used).}
}
\value{
Invisibly returns the original 'disag_model_mmap_mcmc' object.
}
\description{
Displays a brief overview of a multi-map disaggregation model fit via MCMC.
}
