% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.summary.disag_model_mmap_mcmc}
\alias{print.summary.disag_model_mmap_mcmc}
\title{Print method for 'summary.disag_model_mmap_mcmc' objects}
\usage{
\method{print}{summary.disag_model_mmap_mcmc}(x, max_print = 30, ...)
}
\arguments{
\item{x}{A 'summary.disag_model_mmap_mcmc' object.}

\item{max_print}{Maximum number of parameters to print.}

\item{...}{Additional arguments (not used).}
}
\value{
Invisibly returns the original summary object.
}
\description{
Displays model metadata and posterior summaries for an MCMC fit.
}
