% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{validate_prepare_data_inputs}
\alias{validate_prepare_data_inputs}
\title{Validate inputs to prepare_data_mmap()}
\usage{
validate_prepare_data_inputs(
  polygon_shapefile_list,
  covariate_rasters_list,
  aggregation_rasters_list,
  id_var,
  response_var,
  sample_size_var,
  make_mesh,
  categorical_covariate_baselines = NULL
)
}
\arguments{
\item{polygon_shapefile_list}{A list of 'sf' objects.}

\item{covariate_rasters_list}{NULL or a list of 'SpatRaster' objects.}

\item{aggregation_rasters_list}{NULL or a list of 'SpatRaster' objects.}

\item{id_var}{Character of length 1: name of polygon ID column.}

\item{response_var}{Character of length 1: name of response column.}

\item{sample_size_var}{NULL or character of length 1: sample-size column.}

\item{make_mesh}{Logical flag indicating whether to build a mesh.}

\item{categorical_covariate_baselines}{passed from prepare_data_mmap}
}
\value{
Invisibly 'TRUE' if all checks pass; otherwise stops with an error.
}
\description{
Check that all list-arguments are the same length (where required),
that required arguments have the correct type, and that
'id_var', 'response_var', and 'sample_size_var' are valid strings.
}
\keyword{internal}
