% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.disag_model_mmap_aghq}
\alias{print.disag_model_mmap_aghq}
\title{Print method for 'disag_model_mmap_aghq' objects}
\usage{
\method{print}{disag_model_mmap_aghq}(x, ..., max_print = 30)
}
\arguments{
\item{x}{A 'disag_model_mmap_aghq' object.}

\item{...}{Additional arguments (not used).}

\item{max_print}{Maximum number of random effects details to print.}
}
\value{
Invisibly returns the original 'disag_model_mmap_aghq' object.
}
\description{
Displays a brief overview of a multi-map disaggregation model:
model family, link function, and components included.
}
